# swagger.yaml
openapi: 3.0.4
info:
  title: Job Portal API
  description: |-
    API documentation for Job Portal - A comprehensive job search and recruitment platform.

    This API provides endpoints for authentication, user management, and administrative operations.
  version: 1.0.0
  contact:
    email: support@jobportal.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: Local development server
  - url: http://localhost:3000/api
    description: Frontend dev proxy (if applicable)
  - url: https://api.jobportal.com/api
    description: Production server
...
paths:
...
  /user/me/profile:
    put:
      tags:
        - user
      summary: Update user profile
      description: Update profile information for the currently authenticated user. At least one field must be provided.
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: 'John Doe Updated'
              phone: '+1234567890'
              location_id: '123e4567-e89b-12d3-a456-426614174000'
              metadata:
                avatar: 'https://example.com/avatar.jpg'
                bio: 'Updated bio'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
              example:
                message: 'Profile updated'
                profile:
                  id: '123e4567-e89b-12d3-a456-426614174002'
                  user_id: '123e4567-e89b-12d3-a456-426614174000'
                  name: 'John Doe Updated'
                  phone: '+1234567890'
                  location_id: '123e4567-e89b-12d3-a456-426614174000'
                  metadata:
                    avatar: 'https://example.com/avatar.jpg'
                    bio: 'Updated bio'
                  updated_at: '2024-01-01T00:00:00.000Z'
                  created_at: '2024-01-01T00:00:00.000Z'
        '400':
          description: Bad request - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/profiles/{id}/public:
    get:
      tags:
        - user
      summary: Get public profile of candidate (Recruiter only)
      description: |-
        Get public profile information of a candidate by profile ID.
        This endpoint allows recruiters to view candidate profiles for recruitment purposes.
        Sensitive information like phone number and email are excluded from the response.
      operationId: getPublicProfile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Profile ID (UUID)
          required: true
          schema:
            type: string
            format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
      responses:
        '200':
          description: Public profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PublicProfileResponse'
              example:
                success: true
                message: 'Profile retrieved successfully'
                data:
                  id: '123e4567-e89b-12d3-a456-426614174000'
                  full_name: 'John Doe'
                  display_name: 'John'
                  headline: 'Senior Software Engineer'
                  avatar_url: 'https://example.com/avatar.jpg'
                  bio: 'Experienced software developer'
                  years_of_experience: 5
                  desired_job_title: 'Full Stack Developer'
                  desired_salary_min: 20000000
                  desired_currency: 'VND'
                  desired_job_type: ['full_time', 'remote']
                  is_looking_for_job: true
                  location_text: 'Ho Chi Minh City'
                  personal_website: 'https://johndoe.dev'
                  linkedin_url: 'https://linkedin.com/in/johndoe'
                  github_url: 'https://github.com/johndoe'
                  role: 'candidate'
                  location:
                    id: '123e4567-e89b-12d3-a456-426614174001'
                    name: 'Ho Chi Minh City'
                    type: 'city'
                  experiences: []
                  educations: []
                  skills: []
                  certifications: []
                  awards: []
        '401':
          description: Unauthorized - Invalid or missing access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Recruiter access required or profile not available for public viewing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: 'This profile is not available for public viewing'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: 'Profile not found'
        '422':
          description: Unprocessable Entity - Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================
  # ADMIN ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT token obtained from login or register endpoint. Format: "Bearer {token}"'
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: 'Operation successful'
      required:
        - statusCode
        - message
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP status code
        message:
          type: string
          example: 'Validation failed'
          description: Error message describing what went wrong
        error:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            'body.email': ['Email must be valid']
            'body.password': ['Password is required']
          description: Detailed validation errors (optional, only present for validation failures)
      required:
        - statusCode
        - message

    # ============================================
    # Enums
    # ============================================
...
    UpdateProfileRequest:
      type: object
      description: At least one field must be provided
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: 'John Doe'
          description: User's full name (1-100 characters)
        phone:
          type: string
          pattern: '^\+?[0-9\s\-().]{7,20}$'
          example: '+1234567890'
          description: Phone number in international format
        location_id:
          type: string
          format: uuid
          nullable: true
          example: '123e4567-e89b-12d3-a456-426614174000'
          description: Location ID (UUID) or null
        metadata:
          type: object
          additionalProperties: true
          example:
            avatar: 'https://example.com/avatar.jpg'
            bio: 'Software developer'
          description: Additional metadata as JSON object

    # ============================================
    # User Response Schemas
    # ============================================
...
    UpdateProfileResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Profile updated'
        profile:
          type: object
          properties:
            id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            name:
              type: string
            phone:
              type: string
            location_id:
              type: string
              format: uuid
              nullable: true
            metadata:
              type: object
              additionalProperties: true
            updated_at:
              type: string
              format: date-time
            created_at:
              type: string
              format: date-time
    PublicProfileResponse:
      type: object
      description: Public profile information for recruiters (excludes sensitive data like phone and email)
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        full_name:
          type: string
          example: 'John Doe'
        display_name:
          type: string
          nullable: true
          example: 'John'
        headline:
          type: string
          nullable: true
          example: 'Senior Software Engineer'
        avatar_url:
          type: string
          nullable: true
          example: 'https://example.com/avatar.jpg'
        bio:
          type: string
          nullable: true
          example: 'Experienced software developer'
        years_of_experience:
          type: integer
          nullable: true
          example: 5
        desired_job_title:
          type: string
          nullable: true
          example: 'Full Stack Developer'
        desired_salary_min:
          type: integer
          nullable: true
          example: 20000000
        desired_currency:
          type: string
          nullable: true
          example: 'VND'
        desired_job_type:
          type: array
          items:
            type: string
          nullable: true
          example: ['full_time', 'remote']
        is_looking_for_job:
          type: boolean
          nullable: true
          example: true
        location_text:
          type: string
          nullable: true
          example: 'Ho Chi Minh City'
        personal_website:
          type: string
          nullable: true
          example: 'https://johndoe.dev'
        linkedin_url:
          type: string
          nullable: true
          example: 'https://linkedin.com/in/johndoe'
        github_url:
          type: string
          nullable: true
          example: 'https://github.com/johndoe'
        role:
          type: string
          example: 'candidate'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        location:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            type:
              type: string
            parent:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                type:
                  type: string
        experiences:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              company_name:
                type: string
              position:
                type: string
              start_date:
                type: string
                format: date-time
              end_date:
                type: string
                format: date-time
                nullable: true
              is_current:
                type: boolean
              description:
                type: string
                nullable: true
        educations:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              school_name:
                type: string
              degree:
                type: string
                nullable: true
              field_of_study:
                type: string
                nullable: true
              start_date:
                type: string
                format: date-time
              end_date:
                type: string
                format: date-time
                nullable: true
        skills:
          type: array
          items:
            type: object
            properties:
              skill_id:
                type: string
                format: uuid
              proficiency:
                type: integer
                nullable: true
              level:
                type: string
                nullable: true
                enum: [beginner, intermediate, advanced, expert]
              skills:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  category:
                    type: string
                    nullable: true
        certifications:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              issuing_org:
                type: string
              credential_id:
                type: string
                nullable: true
              credential_url:
                type: string
                nullable: true
              issue_date:
                type: string
                format: date-time
              expiry_date:
                type: string
                format: date-time
                nullable: true
              never_expires:
                type: boolean
              description:
                type: string
                nullable: true
              skills_acquired:
                type: string
                nullable: true
        awards:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              title:
                type: string
              issuer:
                type: string
              date:
                type: string
                format: date-time
              description:
                type: string
                nullable: true
              url:
                type: string
                nullable: true
              category:
                type: string
                nullable: true
                enum: [academic, professional, competition, volunteer, other]
              level:
                type: string
                nullable: true
                enum: [international, national, regional, local, organizational]

    # ============================================
    # Admin Request Schemas
    # ============================================
...